Lesson 4 refactoring
