+++ app/src/main/java/com/nick_sib/popularlibraries/deleteme/ModelLesson4.kt
@@ -21,19 +21,23 @@ object ModelLesson4 {
 
 
 
-    fun convert(bitmap: Bitmap, imagesRoot: String, fileName: String): ConnectableObservable<Int> =
-         Observable.create<Int>{ emitter ->
-             for (i in 1 until MAX_STEPS) {
-                 Thread.sleep(1000)
-                 emitter.onNext(i)
+    fun convert(imagePath: Uri, imagesRoot: String): ConnectableObservable<Int> =
+
+
+        Observable.create<Int>{ emitter ->
+            for (i in 1 until MAX_STEPS) {
+                Thread.sleep(1000)
+                emitter.onNext(i)
              }
 
              //Непосредственно конвертация
-             val myImagesDir = File(imagesRoot + IMAGE_FOLDER)
+            val myImagesDir = File(imagesRoot + IMAGE_FOLDER)
              if (!myImagesDir.exists()) myImagesDir.mkdirs()
-             val result = String.format("%s%s/%s.png", imagesRoot, IMAGE_FOLDER, fileName)
+             val result = String.format("%s%s/%s.png", imagesRoot, IMAGE_FOLDER, imagePath.getName())
              val out = FileOutputStream(File(result))
-             bitmap.compress(Bitmap.CompressFormat.PNG, 100, out)
+             imagePath
+                 .getCapturedImage()?.compress(Bitmap.CompressFormat.PNG, 100, out)
+                 ?: emitter.onError(Throwable("Image ERROR"))
              emitter.onNext(MAX_STEPS)
              convertedImage = result.toUri()
              out.flush()
@@ -43,6 +47,12 @@ object ModelLesson4 {
             .subscribeOn(Schedulers.io())
             .observeOn(AndroidSchedulers.mainThread())
             .publish()
+
+
+
+
+
+
 
 
 }
